Refactor : Suppression Fichiers inutiles
