--- commentaires/src/main/java/com/ubo/commentairesDAO/DAOException.java
@@ -1,15 +1,0 @@
-package com.ubo.commentairesDAO;
-
-public class DAOException extends java.lang.Exception {
-
-    private static final long serialVersionUID = 309200220164832142L;
-
-    public DAOException() {
-        super();
-    }
-
-    public DAOException(String message) {
-        super(message);
-    }
-
-}
