--- commentaires/src/main/java/com/ubo/commentairesDAO/PersistenceKind.java
@@ -1,10 +1,0 @@
-package com.ubo.commentairesDAO;
-
-/**
- * Type de support de persistance pour les données.
- *
- * @author Maxime A.
- */
-public enum PersistenceKind {
-    JDBC, JPA, XML, JSON
-}
